+++ src/FastLED.h
@@ -137,7 +137,7 @@ enum ESPIChipsets {
 	APA102HD, ///< APA102 LED chipset with 5-bit gamma correction
 	HD107,  /// Same as APA102, but in turbo 40-mhz mode.
 	HD107HD,  /// Same as APA102HD, but in turbo 40-mhz mode.
-	HD108,  /// 16bit variant of HD107, always gamma corrected.
+	HD108,  /// 16-bit variant of HD107, always gamma corrected. No SD (standard definition) option available - all HD108s use gamma correction, and a non-gamma corrected version is not planned.
 
 };